# backend/docs/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Club Booking SaaS Platform API
  description: >
    Multi-tenant booking engine for Players and Club Owners.
    
    ## Authentication
    All protected endpoints use JWT stored in an **HttpOnly cookie** named `access_token_cookie`.
    The browser automatically sends the cookie with each request – no manual token management is required.
    Endpoints that require specific roles (e.g., `owner`, `player`) are marked accordingly.

    ## Implementation Deviations Logged
    1. Operating hours fields (`open_time`, `close_time`, `slot_duration_minutes`) are flat fields on `Club` and `Court` models, not nested objects.
    2. Overlap validation for Bookings is performed at the application layer via Python queries instead of PostgreSQL GiST constraints, as SQLite is used.
    3. Additional modules implemented and documented: Memberships, Events, Payments, Attendance, and Notifications.
  version: 2.1.0
servers:
  - url: http://127.0.0.1:5000/api/v1
    description: Local Flask Development Server
security:
  - cookieAuth: []   # JWT stored in HttpOnly cookie
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
      description: JWT authentication cookie (HttpOnly)
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
        club_id:
          type: integer
        created_by:
          type: integer
        title:
          type: string
        description:
          type: string
        event_date:
          type: string
          format: date
        start_time:
          type: string
          format: time
        end_time:
          type: string
          format: time
        max_attendees:
          type: integer
          nullable: true
        registration_fee:
          type: number
          format: float
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
...
paths:
...
  /bookings:
    post:
      summary: Create a booking (Player Only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [court_id, booking_date, start_time, end_time]
              properties:
                court_id: { type: integer }
                booking_date: { type: string, format: date }
                start_time: { type: string, format: time }
                end_time: { type: string, format: time }
      responses:
        '201':
          description: Booking created successfully
        '400':
          description: Validation error
        '404':
          description: Court not found or inactive
        '409':
          description: Slot already booked

    get:
      summary: Get current user's bookings (Player Only)
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer }
                    court_id: { type: integer }
                    court_name: { type: string }
                    club_name: { type: string }
                    booking_date: { type: string, format: date }
                    start_time: { type: string, format: time }
                    end_time: { type: string, format: time }
                    status: { type: string, enum: [active, released, overridden] }
...
  /events:
    get:
      summary: List upcoming events (public)
      parameters:
        - name: club_id
          in: query
          schema: { type: integer }
          description: Filter by club
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      summary: Create a new event (Owner Only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [club_id, title, event_date, start_time, end_time]
              properties:
                club_id: { type: integer }
                title: { type: string }
                description: { type: string }
                event_date: { type: string, format: date }
                start_time: { type: string, format: time }
                end_time: { type: string, format: time }
                max_attendees: { type: integer, nullable: true }
                registration_fee: { type: number, format: float, default: 0 }
      responses:
        '201':
          description: Event created successfully
        '400':
          description: Validation error
        '403':
          description: Not authorized to create events for this club
        '404':
          description: Club not found
...
  /attendance/check-in:
    post:
      summary: Check in a user (JWT required – any role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              oneOf:
                - required: [booking_id]
                - required: [event_id]
              properties:
                booking_id: { type: integer, nullable: true }
                event_id: { type: integer, nullable: true }
      responses:
        '201':
          description: Checked in successfully
        '400':
          description: Validation error
        '404':
          description: Booking/event not found
        '409':
          description: Already checked in
...